Add file-transfer service
The file-transfer sevice will be used to send/receive files between peers and will form the base for more complex download tasks. This service runs in a separate child process and handles all files being sent to or received from a peer. The changes in each module are as follows:

file-transfer (new): This is the service which handles data transfer between peers and keeps track of files currently being sent/received. It also handles inter process communication to/from the main process.

subprocess: Add a name for the file-transfer child process.

peer-connection: Add requestFileDownload method which sends a message to a peer requesting it to send a file. It also notifies the file-transfer process about the request.

client-connection: Add new messageHandler condition to handle file download requests. It notifes the file-transfer child process about the request.

file-transfer-message (new): FileTransferMessage extends the Message class and adds functionality specific to the file-transfer packets. This class encodes message data into a buffer object to allow for more data to be sent in each packet by reducing wasted bytes in the message.

file-send-request (new): Each FileSendRequest object handles a file currently being sent to a peer. It reads 'chunks' of the file based on the chunks being requested by the peer and sends them (one in each packet) to the peer.

file-receive-request (new): Each FileReceiveRequest handles a file currently being received from a peer. It sends packets to the peer containing requests for specific chunks of the file, and writes these chunks to the file (in disk) when it receives them from the peer.

renderer: Start the file-transfer child process when the application starts. Also ensure that the child process is killed before the application exits.

The 'defaultDownloadPath' field is added to settings. Currently all received files are saved to this path.

There are still sereval performance improvement and code cleanup tasks that need to be done. These will be added incrementally over several future commits.